CIRC-626: Remove claim-item-returned property class. Move claimItemReturned method to Loan.
