--- src/main/java/org/folio/circulation/domain/representations/ClaimItemReturnedProperties.java
@@ -1,9 +1,0 @@
-package org.folio.circulation.domain.representations;
-
-public class ClaimItemReturnedProperties {
-
-  private ClaimItemReturnedProperties() {}
-
-  public static final String ITEM_CLAIMED_RETURNED_DATE = "itemClaimedReturnedDateTime";
-  public static final String COMMENT = "comment";
-}
